Extract the activity record validity checking into a separate class to reduce the LOC in the WorkspaceActivityManager.
