+++ multiuser/api/che-multiuser-api-workspace-activity/src/test/java/org/eclipse/che/multiuser/api/workspace/activity/MultiUserWorkspaceActivityManagerTest.java
@@ -62,7 +62,6 @@ public class MultiUserWorkspaceActivityManagerTest {
     activityManager =
         new MultiUserWorkspaceActivityManager(
             workspaceManager,
-            workspaceRuntimes,
             workspaceActivityDao,
             eventService,
             accountManager,
